ADD ANALYTICAL REPORTS
Reason:
- Add analytical reports about clients
- Add analytical report about MRR value

This commit will add analytical reports regarding the
data exploration about the transactional data about
the plans acquisition by clients during the period between
September/2016 and December/2019.

My first idea was to analyze the MRR variation by means of seasonality
and trend, but it seems like there is no apparent seasonality or trend
in the MRR variation.

Following this observation my next steps will be to analyze other metrics,
including the MRR again but this time separating the contributions of distinct
Market Segments and Regions.